Fixes 0

+++ app/Medians/Media/Application/VideoController.php
@@ -450,10 +450,12 @@ class VideoController extends CustomController
         $response = curl_exec($ch);
 
         $save = file_put_contents($tempFileFullPath, $response);
-
-        $filesize = filesize($tempFileFullPath);
-        $filesize < 100 ? unlink($tempFileFullPath)   : null;
-        return $filesize < 100 ? throw new \Exception("File size is ".$filesize, 1) : true;
+        if ($save)
+        {
+            $filesize = filesize($tempFileFullPath);
+            $filesize < 100 ? unlink($tempFileFullPath)   : null;
+            return $filesize < 100 ? throw new \Exception("File size is ".$filesize, 1) : true;
+        }
     }
